优化

+++ mvvmhabit/src/main/java/me/goldze/mvvmhabit/http/net/entity/RecommendFriendEntity.java
@@ -126,7 +126,7 @@ public class RecommendFriendEntity extends BaseEntity{
         private Object mail;
         private int age;
         private int buildNum;
-        private int cumulativeLoginTime;
+        private long cumulativeLoginTime;
         private Object wxappProvince;
         private int enable;
         private Object wechatOpenId;
@@ -506,11 +506,11 @@ public class RecommendFriendEntity extends BaseEntity{
             this.buildNum = buildNum;
         }
 
-        public int getCumulativeLoginTime() {
+        public long getCumulativeLoginTime() {
             return cumulativeLoginTime;
         }
 
-        public void setCumulativeLoginTime(int cumulativeLoginTime) {
+        public void setCumulativeLoginTime(long cumulativeLoginTime) {
             this.cumulativeLoginTime = cumulativeLoginTime;
         }
